config: Move save_system_paths to system_settings.BpodDir
